Refactor directory structure in tools
Refactored directory structure to make it more understandable and
remove unnecessary python import path editing.  All python modules
are now subpackages of the pymomo super package. References to
these packages have been updated in their dependences but some
lingering errors may remain that will be cleaned up as they are
hit.
