+++ tools/bin/create_skeleton.py
