+++ tools/bin/dump_fat.py
