+++ tools/pymomo/gpysim/statements/logassert.py
