+++ tools/pymomo/gpysim/statements/logfinished.py
