+++ tools/pymomo/hex8/decode.py
