+++ tools/pymomo/hex8/symbols.py
