+++ tools/pymomo/mib/handler.py
@@ -1,7 +1,6 @@
 import sys
 import os.path
-sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
-from momo_utilities import build
+from pymomo.utilities import build
 
 class MIBHandler:
 	"""
