+++ tools/pymomo/pybom/reference.py
@@ -7,9 +7,7 @@ import os.path
 import re
 import sys
 
-sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
-
-from momo_utilities.config import ConfigFile
+from ..utilities.config import ConfigFile
 
 data_file = ConfigFile('pcb_library')
