--- tools/python_modules/mib/test.mib
@@ -1,5 +1,0 @@
-feature 0
-hello(1, yes);
-1: goodbye(2, no);
- #hello
-Feature 1
