+++ tools/site_scons/site_tools/gpsim_runner.py
@@ -4,10 +4,10 @@ import SCons.Builder
 import os
 import sys
 
-sys.path.append(os.path.join(os.path.dirname(__file__), '..', '..', 'python_modules'))
-import momo_utilities.config
+sys.path.append(os.path.join(os.path.dirname(__file__), '..', '..'))
+from pymomo.utilities import config
 
-settings = momo_utilities.config.ConfigFile('settings')
+settings = config.ConfigFile('settings')
 gpsim_path = settings['external_tools/gpsim']
 
 run_gpsim = SCons.Builder.Builder(
